Standardization and fix of systemd
We don't need manual symlinks creation for ucarp anymore.
CentOS 7.3 does not need and also does not work well for this workaround
that we had for systemd. So removing it, as advised by @ramereth

Newly available platforms added to kitchen.
CentOS added to kitchen and upgraded versions
Bringing kitchen.yml near our current ways of using TK

Locking down dependency on 1.0.2 of yum-epel cookbook since 2.0.0 results
in an issue for which no workaround was found.
Reported at https://github.com/chef-cookbooks/yum-epel/issues/41

Spec for networking and template service notifying it

apt is a requriement for debian platform
Service defintion already exists in the data_bag recipe for ucarp so we are
removing it from default recipe

Depend on newer version of yum-epel

Reverting the version bump for the cookbook.
It is handled automatically in our case by manatee and CI so we don't need to
do it

Specs for databag recipe
Stubs and general OS attributes

default recipe and spec should do nothing with ucarp service

The service is handled by data_bag recipe after all.

Also, checking in standard Rakefile
Created by using chef generate cookbook with generator at
/opt/code_generator

ServerSpec Tests fixed to make them reliable

Unit tests for this cookbook don't pass because our version of fauxhai does not
support CentOS 7.3 (just like the upstream) and is also missing support for
6.8 (unlike the upstream) hence we are disabling them for now.

We have decided to rely on ServerSpecs for this cookbook and
get all changes merged

Also, styling fixes